% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_anomaly.R
\name{sst_anomaly}
\alias{sst_anomaly}
\title{Compute monthly grid cell specific SST anomalies}
\usage{
sst_anomaly(data, ref.years)
}
\arguments{
\item{data}{list of SST data output from \code{\link{sst_load}}}

\item{ref.years}{years over which to calculate the long-term average,
years must be contained in input data}
}
\value{
A list with elements:
    \itemize{
       \item{lon: vector of longitudes}
       \item{lat: vector of latitudes}
       \item{year: vector of years}
       \item{month: vector of months}
       \item{sst: 3D array of SST anomalies}
       \item{version: numeric, version of SST data}
     }
}
\description{
Anomalies are calculated as the difference between a grid cell specific
     SST value for a given year/month and the long-term monthly mean (defined
     by ref.years) for that grid cell. This follows the methods outlined in
     Mueter et al. 2002, Canadian Journal of Fisheries and Aquatic Sciences.
}
\examples{
\dontrun{
     sst_download(1950, 1:5, "./sst/", version = 5)
     x <- sst_load(1950, 1:5, "./sst/", version = 5)
     xa <- sst_anomaly(x, ref.years = 1950)
}
}
\seealso{
\code{\link{sst_load}}
}
\author{
Michael Malick
}
