% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_dataframe.R
\name{sst_dataframe}
\alias{sst_dataframe}
\title{Convert SST data array to a data.frame}
\usage{
sst_dataframe(data, na.rm = FALSE)
}
\arguments{
\item{data}{list of SST data output from \code{\link{sst_load}}}

\item{na.rm}{logical, should cells with NA sst values be removed}
}
\value{
A data.frame with columns:
    \itemize{
       \item{year: vector of years}
       \item{month: vector of months}
       \item{lon: vector of longitudes}
       \item{lat: vector of latitudes}
       \item{id: unique grid cell id}
       \item{sst: sst data}
     }
}
\description{
Takes input SST data output from \code{\link{sst_load}} and spatially
     subsets the array.
}
\examples{
\dontrun{
     sst_download(1950, 1:5, "./sst/", version = 5)
     x <- sst_load(1950, 1:5, "./sst/", version = 5)
     xdf <- sst_dataframe(x)
     ndf <- sst_dataframe(x, na.rm = TRUE)

}
}
\seealso{
\code{\link{sst_load}}
}
\author{
Michael Malick
}
