% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_load.R
\name{sst_load}
\alias{sst_load}
\title{Load downloaded ERSST data into R}
\usage{
sst_load(years, months, read.dir, version)
}
\arguments{
\item{years}{vector of years to download, all years should be > 1983, should be a
numeric sequence, ascending by 1}

\item{months}{vector of months to download (integers 1:12), should be a
numeric sequence, ascending by 1}

\item{read.dir}{path to directory to read sst data}

\item{version}{numeric, version of SST data to download}
}
\value{
A list with elements:
    \itemize{
       \item{lon: vector of longitudes}
       \item{lat: vector of latitudes}
       \item{year: vector of years}
       \item{month: vector of months}
       \item{sst: 3D array of SST anomalies}
       \item{version: numeric, version of SST data}
     }

     The SST array has dimensions: lon (rows) x lat (columns) x month/year
     (slices)

     Each matrix/slice in the array is gridded SST data for a given month and
     year, which can be indexed using the $year and $month vectors in the
     list.
}
\description{
Load SST data in netCDF file format into R
}
\details{
Requires the \code{ncdf4} package to read data
}
\examples{
\dontrun{
     sst_download(1950, 1:5, "./sst/", version = 5)
     x <- sst_load(1950, 1:5, "./sst/", version = 5)
     filled.contour(x = x$lon, y = x$lat, z = x$sst[ , , 1])
}
}
\seealso{
\code{\link{sst_download}}
}
\author{
Michael Malick
}
