% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_subset_time.R
\name{sst_subset_time}
\alias{sst_subset_time}
\title{Temporally subset SST data}
\usage{
sst_subset_time(data, years, months)
}
\arguments{
\item{data}{list of SST data output from \code{\link{sst_load}}}

\item{years}{vector of years to subset}

\item{months}{vector of months to subset (integers 1:12)}
}
\value{
A list with elements:
    \itemize{
       \item{lon: vector of longitudes}
       \item{lat: vector of latitudes}
       \item{year: vector of years}
       \item{month: vector of months}
       \item{sst: 3D array of SST anomalies}
       \item{version: numeric, version of SST data}
     }
}
\description{
Takes input SST data output from \code{\link{sst_load}} and temporally
     subsets the array.
}
\examples{
\dontrun{
     sst_download(1950, 1:5, "./sst/", version = 5)
     x <- sst_load(1950, 1:5, "./sst/", version = 5)
     xs <- sst_subset_time(x, 1950, 3)
     filled.contour(x = xs$lon, y = xs$lat, z = xs$sst[ , , 1])
}
}
\seealso{
\code{\link{sst_load}}
}
\author{
Michael Malick
}
